+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Improvements:
 
   - add support for injecting dot evn files via `__PA_SETTINGS_DOT_ENVS`
+  - automatically load watchfiles if dependency available
   
 # 1.2.0